runtime: fix operand coercion errors silently vanishing in operator evaluation
Evaluate() filtered out any operand whose let-coercion failed instead of
short-circuiting with the coercion's own error, so the real error (e.g.
Overflow) never reached the caller and EvaluateExpressionResult indexed
into a shrunk operand array instead, throwing IndexOutOfRangeException.
Affects every operator family, not just one.
