+++ RDCore.Tests/Semantics/Runtime/OperatorLogicalRuntimeSemanticsTests.cs
@@ -21,8 +21,10 @@ namespace RDCore.Tests.Semantics.Runtime;
 /// </summary>
 public abstract class OperatorLogicalRuntimeSemanticsTests : OperatorArithmeticRuntimeSemanticsTests
 {
+    // Identity must be a real SyntaxNodeId (see OperatorLetCoerceRuntimeSemanticsTests for why): the
+    // real coercion provider's recursion guard hashes LetCoercionStackFrame, which hashes this Identity.
     private static readonly VBBinaryOperatorExpressionNode ThrowawayBinary = new(
-        "And", default, TestLocations.TestLocation,
+        "And", NodeId, TestLocations.TestLocation,
         [
             new LiteralExpressionNode(default, TestLocations.TestLocationLHS, new VBLongValue(0)),
             new LiteralExpressionNode(default, TestLocations.TestLocationRHS, new VBLongValue(0)),
